move protected table data and column to custom hook
